requirements.txt
openai>=1.14.3
pyyaml>=6.0
pytest>=7.4
